+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [3.9.1] 2019-10-25
+
+### Added
+
+- [TD-2187] StructureCache: force option in put function.
+
+### Changed
+
+- [TD-2187] StructureCache: store Structure external_id and updated_at.
+
 ## [3.9.0] 2019-10-22
 
 ### Added